requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
APScheduler==3.10.4
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.3
Django==5.2.5
django-apscheduler==0.6.2
django-filter==24.3
djangorestframework==3.15.2
easydict==1.13
flatbuffers==25.2.10
fonttools==4.59.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
joblib==1.5.1
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-ml-py==12.575.51
onnx==1.18.0
onnxruntime-gpu==1.22.0
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pip-check==3.1
prettytable==3.16.0
protobuf==6.32.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
simsimd==6.5.1
six==1.17.0
sqlparse==0.5.3
stringzilla==3.12.6
sympy==1.14.0
terminaltables==3.1.10
threadpoolctl==3.6.0
tifffile==2025.6.11
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.13
